break encoders and decoders into multiple files
