+++ tests/test_codec_opus_encoder.py
@@ -5,7 +5,7 @@ import pytest
 from tests.assert_helper import get_test_example_file, get_test_examples_directory
 
 from facefusion import state_manager
-from facefusion.codecs.opus import create_opus_decoder, create_opus_encoder, decode_opus_buffer, destroy_opus_decoder, destroy_opus_encoder, encode_opus_buffer
+from facefusion.codecs.opus_encoder import create_opus_encoder, destroy_opus_encoder, encode
 from facefusion.common_helper import is_linux, is_macos, is_windows
 from facefusion.download import conditional_download
 from facefusion.ffmpeg import read_audio_buffer
@@ -27,13 +27,13 @@ def test_create_opus_encoder() -> None:
 	assert create_opus_encoder(0, 0) is None
 
 
-def test_encode_opus_buffer() -> None:
+def test_encode() -> None:
 	audio_buffer = read_audio_buffer(get_test_example_file('source.mp3'), 48000, 16, 2)
 	audio_sample = numpy.frombuffer(audio_buffer, dtype = numpy.int16).astype(numpy.float32) / 32768.0
 	opus_encoder = create_opus_encoder(48000, 2)
 
 	if is_linux() or is_windows():
-		assert create_hash(encode_opus_buffer(opus_encoder, audio_sample.tobytes(), 960)) == '8abe71cf'
+		assert create_hash(encode(opus_encoder, audio_sample.tobytes(), 960)) == '8abe71cf'
 
 	if is_macos():
 		pytest.skip()
@@ -45,29 +45,3 @@ def test_destroy_opus_encoder() -> None:
 	with patch.object(opus_module.create_static_library(), 'opus_encoder_destroy') as mock:
 		destroy_opus_encoder(opus_encoder)
 		mock.assert_called_once_with(opus_encoder)
-
-
-#TODO: needs review
-def test_create_opus_decoder() -> None:
-	assert create_opus_decoder(48000, 2)
-	assert create_opus_decoder(0, 0) is None
-
-
-#TODO: needs review
-def test_decode_opus_buffer() -> None:
-	audio_buffer = read_audio_buffer(get_test_example_file('source.mp3'), 48000, 16, 2)
-	audio_sample = numpy.frombuffer(audio_buffer, dtype = numpy.int16).astype(numpy.float32) / 32768.0
-	opus_encoder = create_opus_encoder(48000, 2)
-	encoded_buffer = encode_opus_buffer(opus_encoder, audio_sample.tobytes(), 960)
-	opus_decoder = create_opus_decoder(48000, 2)
-
-	assert len(decode_opus_buffer(opus_decoder, encoded_buffer, 960, 2)) == 960 * 2 * 4
-
-
-#TODO: needs review
-def test_destroy_opus_decoder() -> None:
-	opus_decoder = create_opus_decoder(48000, 2)
-
-	with patch.object(opus_module.create_static_library(), 'opus_decoder_destroy') as mock:
-		destroy_opus_decoder(opus_decoder)
-		mock.assert_called_once_with(opus_decoder)
